Move each class to its own file
