+++ app/controllers/api/v2/children_controller.rb
@@ -1,0 +1,7 @@
+module Api
+    module V2
+      class ChildrenController < JSONAPI::ResourceController
+      end
+    end
+  end
+
