+++ app/controllers/api/v2/downstream_assets_controller.rb
@@ -1,0 +1,7 @@
+module Api
+    module V2
+      class DownstreamAssetsController < JSONAPI::ResourceController
+      end
+    end
+  end
+
