+++ app/controllers/api/v2/parents_controller.rb
@@ -1,0 +1,6 @@
+module Api
+  module V2
+    class ParentsController < JSONAPI::ResourceController
+    end
+  end
+end
