+++ app/controllers/api/v2/qc_results_controller.rb
@@ -11,8 +11,6 @@ module Api
           @qc_result_factory.save
           @qc_result_resources = @qc_result_factory.qc_results.map { |qc_result| QcResultResource.new(qc_result, nil) }
 
-          #render json: JSONAPI::ResourceSerializer.new(QcResultResource).serialize_to_hash(@qc_result_resources),
-          #       status: :created
           render json: serialize_array(@qc_result_resources), status: :created
         else
           render json: @qc_result_factory.errors, status: :unprocessable_entity
