+++ app/controllers/api/v2/source_receptacles_controller.rb
@@ -1,0 +1,6 @@
+module Api
+  module V2
+    class SourceReceptaclesController < JSONAPI::ResourceController
+    end
+  end
+end
