+++ app/controllers/api/v2/upstream_assets_controller.rb
@@ -1,0 +1,7 @@
+module Api
+    module V2
+      class UpstreamAssetsController < JSONAPI::ResourceController
+      end
+    end
+  end
+
