+++ app/resources/api/v2/child_resource.rb
@@ -1,0 +1,7 @@
+module Api
+    module V2
+      class ChildResource < JSONAPI::Resource
+      end
+    end
+  end
+
