+++ app/resources/api/v2/descendant_resource.rb
@@ -1,0 +1,6 @@
+module Api
+  module V2
+    class DescendantResource < JSONAPI::Resource
+    end
+  end
+end
