+++ app/resources/api/v2/source_receptacle_resource.rb
@@ -1,0 +1,7 @@
+module Api
+    module V2
+      class SourceReceptacleResource < JSONAPI::Resource
+      end
+    end
+  end
+
