+++ config/initializers/patch_json_api_resource.rb
@@ -20,75 +20,77 @@
 #   Rails.logger.warn '*' * 80
 # end
 
-# # Modified from: jsonapi-resources-0.9.0/lib/jsonapi/resource_serializer.rb
-# module JSONAPI
-#   # Disable cops to prevent auto-correct-induced drift
-#
-#   # Reopen ResourceSerializer to fix the polymorphic associations
-#   class ResourceSerializer
-#     def to_many_linkage(source, relationship)
-#       linkage = []
+# Modified from: jsonapi-resources-0.9.0/lib/jsonapi/resource_serializer.rb
+module JSONAPI
+  # Disable cops to prevent auto-correct-induced drift
 
-#       linkage_types_and_values =
-#         if source.preloaded_fragments.key?(format_key(relationship.name))
-#           source.preloaded_fragments[format_key(relationship.name)].map do |_, resource|
-#             [relationship.type, resource.id]
-#           end
-#         elsif relationship.polymorphic?
-#           assoc = source._model.public_send(relationship.name)
+  # Reopen ResourceSerializer to fix the polymorphic associations
+  class ResourceSerializer
 
-#           # Avoid hitting the database again for values already pre-loaded
-#           # MODIFICATION BEGINS
-#           if assoc.respond_to?(:loaded?) && assoc.loaded?
-#             assoc.map { |obj| [source.class.resource_type_for(obj)&.pluralize, obj.id] }
-#           else
-#             type_column = assoc.inheritance_column
-#             assoc
-#               .pluck(type_column, :id)
-#               .map do |type, id|
-#                 [source.class._model_hints[type.underscore]&.pluralize || type.underscore.pluralize, id]
-#               end
-#           end
-#           # MODIFICATION ENDS
-#         else
-#           source.public_send(relationship.name).map { |value| [relationship.type, value.id] }
-#         end
+    # def to_many_linkage(source, relationship)
+    #   linkage = []
 
-#       linkage_types_and_values.each do |type, value|
-#         linkage.append(type: format_key(type), id: @id_formatter.format(value)) if type && value
-#       end
-#       linkage
-#     end
+    #   linkage_types_and_values =
+    #     if source.preloaded_fragments.key?(format_key(relationship.name))
+    #       source.preloaded_fragments[format_key(relationship.name)].map do |_, resource|
+    #         [relationship.type, resource.id]
+    #       end
+    #     elsif relationship.polymorphic?
+    #       assoc = source._model.public_send(relationship.name)
 
-#     def foreign_key_types_and_values(source, relationship)
-#       if relationship.is_a?(JSONAPI::Relationship::ToMany)
-#         if relationship.polymorphic?
-#           assoc = source._model.public_send(relationship.name)
+    #       # Avoid hitting the database again for values already pre-loaded
+    #       # MODIFICATION BEGINS
+    #       if assoc.respond_to?(:loaded?) && assoc.loaded?
+    #         assoc.map { |obj| [source.class.resource_type_for(obj)&.pluralize, obj.id] }
+    #       else
+    #         type_column = assoc.inheritance_column
+    #         assoc
+    #           .pluck(type_column, :id)
+    #           .map do |type, id|
+    #             [source.class._model_hints[type.underscore]&.pluralize || type.underscore.pluralize, id]
+    #           end
+    #       end
+    #       # MODIFICATION ENDS
+    #     else
+    #       source.public_send(relationship.name).map { |value| [relationship.type, value.id] }
+    #     end
 
-#           # Avoid hitting the database again for values already pre-loaded
-#           # MODIFICATION BEGINS
-#           if assoc.respond_to?(:loaded?) && assoc.loaded?
-#             assoc.map { |obj| [source.class.resource_type_for(obj), @id_formatter.format(obj.id)] }
-#           else
-#             type_column = assoc.inheritance_column
-#             assoc
-#               .pluck(type_column, :id)
-#               .map do |type, id|
-#                 [
-#                   source.class._model_hints[type.underscore]&.pluralize || type.underscore.pluralize,
-#                   @id_formatter.format(id)
-#                 ]
-#               end
-#             # MODIFICATION ENDS
-#           end
-#         else
-#           source.public_send(relationship.name).map { |value| [relationship.type, @id_formatter.format(value.id)] }
-#         end
-#       end
-#     end
-#     # rubocop:enable all
-#   end
-# end
+    #   linkage_types_and_values.each do |type, value|
+    #     linkage.append(type: format_key(type), id: @id_formatter.format(value)) if type && value
+    #   end
+    #   linkage
+    # end
+
+    def foreign_key_types_and_values(source, relationship)
+      binding.pry
+      if relationship.is_a?(JSONAPI::Relationship::ToMany)
+        if relationship.polymorphic?
+          assoc = source._model.public_send(relationship.name)
+
+          # Avoid hitting the database again for values already pre-loaded
+          # MODIFICATION BEGINS
+          if assoc.respond_to?(:loaded?) && assoc.loaded?
+            assoc.map { |obj| [source.class.resource_type_for(obj), @id_formatter.format(obj.id)] }
+          else
+            type_column = assoc.inheritance_column
+            assoc
+              .pluck(type_column, :id)
+              .map do |type, id|
+                [
+                  source.class._model_hints[type.underscore]&.pluralize || type.underscore.pluralize,
+                  @id_formatter.format(id)
+                ]
+              end
+            # MODIFICATION ENDS
+          end
+        else
+          source.public_send(relationship.name).map { |value| [relationship.type, @id_formatter.format(value.id)] }
+        end
+      end
+    end
+    # rubocop:enable all
+  end
+end
 
 # Fix: "labware"."id" AS "labware_id" not valid quoting for mysql.
 # TODO: JSON API RESOURCES Version 11 should solve it <https://github.com/cerebris/jsonapi-resources/issues/1369>
